📁 types(patch): improve typings (@roots/bud-react)
