+++ sources/@roots/bud-react/test/babel-refresh/extension.test.ts
@@ -1,6 +1,5 @@
+import Extension from '@roots/bud-react/babel-refresh'
 import {describe, expect, it} from 'vitest'
-
-import Extension from '../../src/index.js'
 
 describe(`@roots/bud-react/babel-refresh`, () => {
   it(`should be constructable`, () => {
