+++ sources/@roots/bud-react/test/extension/extension.test.ts
@@ -1,8 +1,6 @@
-import '@roots/bud-react/types'
-
+import '@roots/bud-react'
+import Extension from '@roots/bud-react/extension'
 import {describe, expect, it} from 'vitest'
-
-import Extension from '../../src/index'
 
 describe(`@roots/bud-react`, () => {
   it(`should be constructable`, () => {
