+++ sources/@roots/bud-react/test/typescript-refresh/extension.test.ts
@@ -1,9 +1,5 @@
-import '@roots/bud-react/types'
-import '@roots/bud-typescript/types'
-
+import Extension from '@roots/bud-react/typescript-refresh'
 import {describe, expect, it} from 'vitest'
-
-import Extension from '../../src/index.js'
 
 describe(`@roots/bud-react/typescript-refresh`, () => {
   it(`should be constructable`, () => {
